revision c10

+++ Challenge_10.c
@@ -2,10 +2,10 @@
 int main()
 {
     float conf;
-    int rayon;
-    int pi=3.14;
+    float rayon;
+    float pi=3.14;
     printf("entrer le rayon : ");
-    scanf("%d",&rayon);
+    scanf("%f",&rayon);
     conf =2*pi*rayon;
     printf("resultat est %f",conf);
